refactor(api-bridge): split commands/types.ts into domain files
Replace the monolithic 1761-line types.ts with a types/ directory
containing 12 focused domain files (roll, actor, chat, journal, combat,
token, item, scene, compendium, table, event, shared) plus base.ts for
the CommandType union, CommandParamsMap, CommandResultMap, and wire
envelope types. A barrel index.ts re-exports everything so all
existing consumer imports from '@/commands/types' remain unchanged.
Zero behavior change; zero new types; zero consumer import changes.
