+++ apps/foundry-api-bridge/src/commands/types/chat.ts
@@ -1,0 +1,15 @@
+/** Chat message command params and results. */
+
+export interface SendChatMessageParams {
+  content: string;
+  speaker?: string;
+  actorId?: string;
+  flavor?: string;
+  whisperTo?: string[];
+  type?: 'ic' | 'ooc' | 'emote';
+}
+
+export interface SendChatMessageResult {
+  messageId: string;
+  sent: boolean;
+}
