+++ apps/foundry-api-bridge/src/commands/types/index.ts
@@ -1,0 +1,21 @@
+/**
+ * Barrel re-export for the commands/types domain.
+ *
+ * Each domain file owns a focused slice of the type surface.
+ * This index re-exports everything so consumer imports remain
+ * unchanged (`from '@/commands/types'`).
+ */
+
+export * from './shared';
+export * from './roll';
+export * from './actor';
+export * from './chat';
+export * from './journal';
+export * from './combat';
+export * from './token';
+export * from './item';
+export * from './scene';
+export * from './compendium';
+export * from './table';
+export * from './event';
+export * from './base';
