+++ apps/foundry-api-bridge/src/commands/types/journal.ts
@@ -1,0 +1,60 @@
+/** Journal entry and journal page command params and results. */
+
+export type JournalPageType = 'text' | 'image' | 'video';
+
+export interface CreateJournalParams {
+  name: string;
+  folder?: string;
+  content?: string;
+  pageType?: JournalPageType;
+}
+
+export interface UpdateJournalParams {
+  journalId: string;
+  name?: string;
+  folder?: string;
+}
+
+export interface DeleteJournalParams {
+  journalId: string;
+}
+
+export interface CreateJournalPageParams {
+  journalId: string;
+  name: string;
+  type?: JournalPageType;
+  content?: string;
+}
+
+export interface UpdateJournalPageParams {
+  journalId: string;
+  pageId: string;
+  name?: string;
+  content?: string;
+}
+
+export interface DeleteJournalPageParams {
+  journalId: string;
+  pageId: string;
+}
+
+// Pull query params
+export type GetJournalsParams = Record<string, never>;
+
+export interface GetJournalParams {
+  journalId: string;
+}
+
+// Journal Results
+export interface JournalPageResult {
+  id: string;
+  name: string;
+  type: string;
+}
+
+export interface JournalResult {
+  id: string;
+  name: string;
+  folder: string | null;
+  pages: JournalPageResult[];
+}
