+++ apps/foundry-api-bridge/src/commands/types/roll.ts
@@ -1,0 +1,58 @@
+/** Dice-rolling params, results, and ability key constants. */
+
+export type AbilityKey = 'str' | 'dex' | 'con' | 'int' | 'wis' | 'cha';
+
+export const ABILITY_KEYS: readonly AbilityKey[] = ['str', 'dex', 'con', 'int', 'wis', 'cha'];
+
+export interface RollDiceParams {
+  formula: string;
+  showInChat?: boolean;
+  flavor?: string;
+}
+
+export interface RollAbilityParams {
+  actorId: string;
+  ability: AbilityKey;
+  showInChat?: boolean;
+}
+
+export interface RollSkillParams {
+  actorId: string;
+  skill: string;
+  showInChat?: boolean;
+}
+
+export interface RollSaveParams {
+  actorId: string;
+  ability: AbilityKey;
+  showInChat?: boolean;
+}
+
+export interface RollAttackParams {
+  actorId: string;
+  itemId: string;
+  advantage?: boolean;
+  disadvantage?: boolean;
+  showInChat?: boolean;
+}
+
+export interface RollDamageParams {
+  actorId: string;
+  itemId: string;
+  critical?: boolean;
+  showInChat?: boolean;
+}
+
+export interface DiceResult {
+  type: string;
+  count: number;
+  results: number[];
+}
+
+export interface RollResult {
+  total: number;
+  formula: string;
+  dice: DiceResult[];
+  isCritical?: boolean;
+  isFumble?: boolean;
+}
